Move changeLogs sync-schema to separate file
